feat: create handleCreateRoom
