+++ .gitignore
@@ -1,10 +1,12 @@
 # dependencies
 /node_modules
 
+
 # System Files
 .DS_Store
 Thumbs.db
 [Dd]esktop.ini
+
 
 # Links
 *.lnk
@@ -14,3 +16,5 @@ Thumbs.db
 
 # Logs
 npm-debug.log
+
+/build
